+++ stdnet/contrib/tasks/queue.py
@@ -1,3 +1,4 @@
+
 from stdnet import orm